# src/main/resources/APIDOC.YAML
openapi: 3.0.1
info:
  title: Constructa OpenApi specification
  description: OpenApi documentation for Constructa BE
  contact:
    name: Ondřej Bureš
    url: 'https://digitalnoodles.cz/'
    email: bureson2@fel.cvut.cz
  version: '1.0'
servers:
  - url: 'http://localhost:8080/'
    description: Local ENV
security:
  - bearerAuth: []
paths:
...
  /api/v1/construction-reports:
    get:
      tags:
        - Construction reports
      operationId: getConstructionReports
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConstructionReportDTO'
      security:
        - bearerAuth: []
    put:
      tags:
        - Construction reports
      operationId: updateConstructionReport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConstructionReportRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ConstructionReportDTO'
      security:
        - bearerAuth: []
    post:
      tags:
        - Construction reports
      operationId: createConstructionReport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConstructionReportRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ConstructionReportDTO'
      security:
        - bearerAuth: []
  /api/v1/companies:
    get:
      tags:
        - Companies
      operationId: getCompanies
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyDTO'
      security:
        - bearerAuth: []
    put:
      tags:
        - Companies
      operationId: editCompany
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CompanyDTO'
      security:
        - bearerAuth: []
    post:
      tags:
        - Companies
      operationId: createCompany
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CompanyDTO'
      security:
        - bearerAuth: []
...
  '/api/v1/vehicles/reports/{vehicleReportId}':
    get:
      tags:
        - Vehicle reports
      operationId: getVehicleReport
      parameters:
        - name: vehicleReportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleReportDTO'
      security:
        - bearerAuth: []
...
  '/api/v1/construction-reports/{constructionReportId}':
    get:
      tags:
        - Construction reports
      operationId: getConstructionReport
      parameters:
        - name: constructionReportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConstructionReportDTO'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Construction reports
      operationId: deleteConstructionReport
      parameters:
        - name: constructionReportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
      security:
        - bearerAuth: []
  '/api/v1/construction-reports/project/{projectId}':
    get:
      tags:
        - Construction reports
      operationId: getConstructionReportsByProjectId
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConstructionReportDTO'
      security:
        - bearerAuth: []
  '/api/v1/companies/{companyId}':
    get:
      tags:
        - Companies
      operationId: getCompanys
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDTO'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Companies
      operationId: deleteCompany
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
      security:
        - bearerAuth: []
  '/api/v1/vehicles/reports/{reportId}':
    delete:
      tags:
        - Vehicle reports
      operationId: deleteVehicleReport
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
      security:
        - bearerAuth: []
components:
  schemas:
...
    LocationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        city:
          type: string
        street:
          type: string
        descriptiveNumber:
          type: string
        name:
          type: string
        country:
          type: string
        postCode:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        titleBeforeName:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        titleAfterName:
          type: string
        bankAccount:
          type: string
        dateOfAcceptance:
          type: string
          format: date-time
        dateOfBirth:
          type: string
          format: date-time
        birthId:
          type: string
        hourRate:
          type: integer
          format: int32
        monthSalary:
          type: integer
          format: int32
        address:
          $ref: '#/components/schemas/LocationDTO'
...
    VehicleDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        factory:
          type: string
        name:
          type: string
        registrationNumber:
          type: string
        conditionMotorcycleWatch:
          type: number
          format: double
        mileage:
          type: number
          format: double
        type:
          type: string
        boughtAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        qrCode:
          type: string
        vinCode:
          type: string
...
    VehicleReportDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        originalConditionMotorcycleWatch:
          type: number
          format: double
        afterworkConditionMotorcycleWatch:
          type: number
          format: double
        cargoMass:
          type: integer
          format: int32
        cargoType:
          type: string
        distance:
          type: number
          format: double
        purchaseOfFuelLitres:
          type: number
          format: double
        timeFrom:
          type: string
          format: date-time
        timeTo:
          type: string
          format: date-time
        description:
          type: string
        vehicle:
          $ref: '#/components/schemas/VehicleDTO'
        driver:
          $ref: '#/components/schemas/UserDTO'
...
    ConstructionReportRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        taskName:
          type: string
        note:
          type: string
        weather:
          type: string
        date:
          type: string
          format: date-time
        executorId:
          type: integer
          format: int64
        state:
          type: string
          enum:
            - FINISHED
            - IN_PROGRESS
            - BLOCKED
        projectId:
          type: integer
          format: int64
    ConstructionReportDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        taskName:
          type: string
        note:
          type: string
        weather:
          type: string
        date:
          type: string
          format: date-time
        executor:
          $ref: '#/components/schemas/UserDTO'
        state:
          type: string
    CompanyRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        din:
          type: string
        cin:
          type: string
        phone:
          type: string
        country:
          type: string
        city:
          type: string
        street:
          type: string
        descriptiveNumber:
          type: string
        postCode:
          type: string
    CompanyDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        din:
          type: string
        cin:
          type: string
        phone:
          type: string
        companyAddress:
          $ref: '#/components/schemas/LocationDTO'
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT auth description
      in: header
      scheme: bearer
      bearerFormat: JWT